+++ spring-cloud-tencent-commons/src/main/java/com/tencent/cloud/common/metadata/MetadataContext.java
@@ -40,6 +40,17 @@ public class MetadataContext {
 	 * transitive context.
 	 */
 	public static final String FRAGMENT_TRANSITIVE = "transitive";
+
+	/**
+	 * disposable Context.
+	 */
+	public static final String FRAGMENT_DISPOSABLE = "disposable";
+
+	/**
+	 * upstream disposable Context.
+	 */
+	public static final String FRAGMENT_UPSTREAM_DISPOSABLE = "upstream-disposable";
+
 	private static final Logger LOG = LoggerFactory.getLogger(MetadataContext.class);
 	/**
 	 * Namespace of local instance.