+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - [Add configurable heartbeat interval support](https://github.com/Tencent/spring-cloud-tencent/pull/444)
 - [feat:enhance Feign and RestTemplate and support Polaris monitor.](https://github.com/Tencent/spring-cloud-tencent/pull/447)
 - [Optimize feign & rest-template circuit-breaker logic](https://github.com/Tencent/spring-cloud-tencent/pull/454)
+- [Feature: Add disposable metadata transfer support](https://github.com/Tencent/spring-cloud-tencent/pull/459)